+++ R/landmarking.R
@@ -24,25 +24,25 @@
 #' @details
 #'  The following features are allowed for this method:
 #'  \describe{
-#'    \item{"decision.stumps"}{Construct a single DT node model induced by the 
+#'    \item{"decision.stumps"}{Construct a single DT node model induced by the
 #'      most informative attribute. The single split (parallel axis) in the data
 #'      has the main goal of establish the linear separability.}
-#'    \item{"elite.nearest.neighbor"}{Select the most informative attributes in 
-#'      the dataset using the information gain ratio to induce the 1-Nearest 
+#'    \item{"elite.nearest.neighbor"}{Select the most informative attributes in
+#'      the dataset using the information gain ratio to induce the 1-Nearest
 #'      Neighbor. With the subset of informative attributes is expected that the
 #'      models induced by 1-Nearest Neighbor should be noise tolerant.}
-#'    \item{"linear.discriminant"}{Apply the Linear Discriminant classifier to 
-#'      construct a linear split (non parallel axis) in the data to establish 
+#'    \item{"linear.discriminant"}{Apply the Linear Discriminant classifier to
+#'      construct a linear split (non parallel axis) in the data to establish
 #'      the linear separability.
-#'    \item{"naive.bayes"}{Evaluate the performance of the Naive Bayes 
-#'      classifier. It assumes that the attributes are independent and each 
+#'    \item{"naive.bayes"}{Evaluate the performance of the Naive Bayes
+#'      classifier. It assumes that the attributes are independent and each
 #'      example belongs to a certain class based on the Bayes probability.}
-#'    \item{"nearest.neighbor"}{This measure evaluate the performance of the 
-#'      1-Nearest Neighbor classifier. It uses the euclidean distance of the 
+#'    \item{"nearest.neighbor"}{This measure evaluate the performance of the
+#'      1-Nearest Neighbor classifier. It uses the euclidean distance of the
 #'      nearest neighbor to determine how noisy is the data.}
-#'    \item{"worst.node"}{Construct a single DT node model induced by the 
-#'      less informative attribute. With the "decision.stumps" measure is 
-#'      possible to define a baseline value of linear separability for a 
+#'    \item{"worst.node"}{Construct a single DT node model induced by the
+#'      less informative attribute. With the "decision.stumps" measure is
+#'      possible to define a baseline value of linear separability for a
 #'      dataset.}
 #'  }
 #' @return Each one of these meta-features generate multiple values (by fold
@@ -214,10 +214,10 @@ worst.node <- function(x, y, split, ...) {
 
 nearest.neighbor <- function(x, y, split, ...) {
 
-  x <- replace.nominal.columns(x)
   aux <- sapply(split, function(test) {
-    prediction <- class::knn(x[-test,], x[test,], y[-test], k=1)
-    accuracy(prediction, y[test])
+    data <- cbind(x, Class=y)
+    prediction <- kknn::kknn(Class~., data[-test,], x[test,], k=1)
+    accuracy(prediction$fitted.values, y[test])
   })
 
   return(aux)
@@ -225,15 +225,16 @@ nearest.neighbor <- function(x, y, split, ...) {
 
 elite.nearest.neighbor <- function(x, y, split, ...) {
 
-  x <- replace.nominal.columns(x)
   aux <- sapply(split, function(test) {
     imp <- dt.importance(x, y, test)
     att <- names(which(imp != 0))
     if(all(imp == 0))
       att <- colnames(x)
-    prediction <- class::knn(x[-test, att, drop=FALSE],
-      x[test, att, drop=FALSE], y[-test], k=1)
-    accuracy(prediction, y[test])
+
+    data <- cbind(x[, att, drop=FALSE], Class=y)
+    prediction <- kknn::kknn(Class ~ ., data[-test,],
+                             x[test, att, drop=FALSE], k=1)
+    accuracy(prediction$fitted.values, y[test])
   })
 
   return(aux)